refactor(commands): extract cross-root range deletion into SelectionDeletion
Moves the self-contained getRootBlockIndex / deleteLeafRange /
deleteCrossRootRange helpers out of Commands.ts into a focused
SelectionDeletion module, leaving Commands.ts closer to a thin entry point.
No behavior change.
